Rename epic files
