+++ src/client/src/ui/spotTile/epics/closePositionEpic.ts
