+++ src/client/src/ui/spotTile/epics/combineSpotTileEpics.ts
@@ -1,8 +1,8 @@
 import { combineEpics } from 'redux-observable'
-import { closePositionEpic } from './closePosition'
+import { closePositionEpic } from './closePositionEpic'
 import { displayCurrencyChartEpic } from './currencyChartEpic'
 import { pricingServiceEpic, pricingHistoryEpic } from './pricingEpics'
-import { publishPriceUpdateEpic } from './publishPrice'
+import { publishPriceUpdateEpic } from './publishPriceUpdateEpic'
 import { spotTileEpic } from './spotTileEpics'
 import { publishTradeExecutedEpic } from './tradeExecutedEpic'
 import { ApplicationDependencies } from 'applicationServices'
